+++ src/Concerns/Comparison.php
@@ -2,39 +2,35 @@
 
 namespace Henzeb\Enumhancer\Concerns;
 
-use BadMethodCallException;
-use Henzeb\Enumhancer\Helpers\EnumGetters;
 use Henzeb\Enumhancer\Helpers\EnumCompare;
 use UnitEnum;
 
 trait Comparison
 {
+    use MagicCalls;
+
     public function equals(UnitEnum|string|int|null ...$equals): bool
     {
         return EnumCompare::equals($this, ...$equals);
     }
 
-    public function __call(string $name, array $arguments): self|bool
+    public function is(UnitEnum|string|int|null $equals): bool
     {
-        if (EnumCompare::isValidCall(self::class, $name, $arguments)) {
-            throw new BadMethodCallException(sprintf('Call to undefined method %s::%s(...)', $this::class, $name));
-        }
+        return $this->equals($equals);
+    }
 
-        $nameIsEnum = !EnumGetters::tryGet(self::class, $name, true);
+    public function isNot(UnitEnum|string|int|null $equals): bool
+    {
+        return !$this->is($equals);
+    }
 
-        if (!$nameIsEnum && method_exists(self::class, '__callStatic')) {
-            return self::__callStatic($name, []);
-        }
+    public function isIn(UnitEnum|string|int|null ...$equals): bool
+    {
+        return $this->equals(...$equals);
+    }
 
-        $value = substr($name, str_starts_with($name, 'isNot') ? 5 : 2);
-
-        if (!EnumGetters::tryGet(self::class, $value, true)) {
-            throw new BadMethodCallException(sprintf('Call to undefined method %s::%s(...)', $this::class, $name));
-        }
-
-        if (str_starts_with($name, 'isNot')) {
-            return !$this->equals($value);
-        }
-        return $this->equals($value);
+    public function isNotIn(UnitEnum|string|int|null ...$equals): bool
+    {
+        return !$this->equals(...$equals);
     }
 }