+++ src/Concerns/Constructor.php
@@ -6,8 +6,5 @@ use Henzeb\Enumhancer\Helpers\EnumGetters;
 
 trait Constructor
 {
-    public static function __callStatic(string $name, array $arguments)
-    {
-        return EnumGetters::get(self::class, $name, true);
-    }
+    use MagicCalls;
 }